Mouse Sensitivity now persists between sessions
Mouse Cursor now VERY FORCE locks to center of the screen and invis when on a scene with walkability and not in a menu (Done at the OS level)

Login Text now states "Welcome" each time the LogIn button is clicked and only overwrites to information if info is needed (Such as "User Does Not Exist" and "Incorrect Username or Password")

ProcedureStatus resetting database side is now fixed.  Checks for greater ONLY

Started work on NG+
It grabs all final score 100 messages for the Comprehensive Test.  Will sort out only getting the most recent
Then needs to check OS Date and see if enough time has passed to constitute a retest prompt